# lib/api-spec/openapi.yaml
openapi: "3.1.0"
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: "1.0.0"
  description: VoxAgent AI Voice Bot Platform API
servers:
  - url: /api
paths:
...
  /v1/calls:
    get:
      operationId: listCalls
      summary: List calls with filters
      parameters:
        - in: query
          name: direction
          schema:
            type: string
            enum: [INBOUND, OUTBOUND]
        - in: query
          name: hangupReason
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
        - in: query
          name: botId
          schema:
            type: string
      responses:
        "200":
          description: List of calls
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CallList"
  /v1/calls/dial:
    post:
      operationId: dialCall
      summary: Initiate an outbound call
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DialInput"
      responses:
        "201":
          description: Call initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Call"
        "400":
          description: Bad request
  /v1/calls/outbound:
    post:
      operationId: outboundCall
      summary: Initiate an outbound call (compatibility alias for dial)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DialInput"
      responses:
        "201":
          description: Call initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Call"
  /v1/calls/inbound:
    post:
      operationId: receiveInboundCall
      summary: Webhook — incoming call arrives; stamps active persona at call-start
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InboundCallInput"
      responses:
        "201":
          description: Inbound call record created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Call"
        "400":
          description: Bad request
  /v1/calls/{id}:
    get:
      operationId: getCall
      summary: Get call details and status
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Call details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Call"
        "404":
          description: Not found
    delete:
      operationId: hangupCall
      summary: Hang up an active call
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Call hung up
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Call"
        "404":
          description: Not found
  /v1/calls/{id}/transfer:
    post:
      operationId: transferCall
      summary: Transfer call to extension or number
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransferInput"
      responses:
        "200":
          description: Transfer initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Call"
        "404":
          description: Not found
...
  /v1/calls/{id}/conference:
    post:
      operationId: conferenceCall
      summary: Add a party to a conference
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConferenceInput"
      responses:
        "200":
          description: Conference created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Call"
        "404":
          description: Not found
  /v1/calls/{id}/media-events:
    post:
      operationId: recordCallMediaEvent
      summary: Persist an authenticated telephony disclosure or recording-consent event
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CallMediaEventInput"
      responses:
        "200":
          description: Updated call
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Call"

  # ─── Compliance ─────────────────────────────────────────────────
...
components:
...
  schemas:
...
    Call:
      type: object
      required: [id, botId, direction, status, createdAt]
      properties:
        id:
          type: string
        botId:
          type: string
        direction:
          type: string
          enum: [INBOUND, OUTBOUND]
        status:
          type: string
          enum: [INITIATING, RINGING, IN_PROGRESS, COMPLETED, FAILED]
        customerNumber:
          type: ["string", "null"]
        customerName:
          type: ["string", "null"]
        startedAt:
          type: ["string", "null"]
          format: date-time
        endedAt:
          type: ["string", "null"]
          format: date-time
        durationSeconds:
          type: ["integer", "null"]
        hangupReason:
          type: ["string", "null"]
        sipCode:
          type: ["integer", "null"]
        amdResult:
          type: ["string", "null"]
        languageDetected:
          type: ["string", "null"]
        recordingUrl:
          type: ["string", "null"]
        summary:
          type: ["string", "null"]
        transferTarget:
          type: ["string", "null"]
        followUpSent:
          type: boolean
        personaId:
          type: ["string", "null"]
          description: ID of the persona that was active when this call started
        composedPrompt:
          type: ["string", "null"]
          description: Fully-composed system prompt injected for this call
        personaName:
          type: ["string", "null"]
        personaVersion:
          type: ["integer", "null"]
        connectOutcome:
          type: ["string", "null"]
        interruptionCount:
          type: integer
        escalationCount:
          type: integer
        languageSwitches:
          type: ["array", "null"]
          items: { type: object, additionalProperties: true }
        finalDisposition:
          type: ["string", "null"]
        disclosureText:
          type: ["string", "null"]
        disclosurePlayedAt:
          type: ["string", "null"]
          format: date-time
        recordingConsentStatus:
          type: ["string", "null"]
        recordingConsentAt:
          type: ["string", "null"]
          format: date-time
        complianceDecisionId:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    CallList:
      type: object
      required: [calls, total]
      properties:
        calls:
          type: array
          items:
            $ref: "#/components/schemas/Call"
        total:
          type: integer
    DialInput:
      type: object
      required: [to, botId]
      properties:
        to:
          type: string
          description: E.164 phone number
        from:
          type: ["string", "null"]
          description: Caller ID (E.164)
        botId:
          type: string
        maxRingSeconds:
          type: integer
          default: 30
        maxAttempts:
          type: integer
          default: 3
    InboundCallInput:
      type: object
      required: [from, botId]
      properties:
        from:
          type: string
          description: Caller's E.164 phone number
        botId:
          type: string
        to:
          type: ["string", "null"]
          description: Dialed number (DID) that received the call
...
    CallMediaEventInput:
      type: object
      required: [event, evidence]
      properties:
        event: { type: string, enum: [DISCLOSURE_PLAYED, RECORDING_CONSENT_GRANTED, RECORDING_CONSENT_DECLINED] }
        evidence: { type: string }
        occurredAt: { type: string, format: date-time }
    TransferInput:
      type: object
      required: [mode]
      properties:
        mode:
          type: string
          enum: [BLIND, ATTENDED]
        extension:
          type: ["string", "null"]
        e164:
          type: ["string", "null"]
        agentName:
          type: ["string", "null"]
        announceMessage:
          type: ["string", "null"]
        waitTimeoutSeconds:
          type: integer
          default: 20
    ConferenceInput:
      type: object
      required: [target]
      properties:
        target:
          type: string
          description: E.164 number or extension to add
        dropBotAfterConnect:
          type: boolean
          default: false

    # ─── Stats schemas ────────────────────────────────────────────
...